# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Node.js REPL API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ExecuteInput {
  /** JavaScript/Node.js code to execute */
  code: string;
  /**
     * Execution timeout in milliseconds (default 10000)
     * @nullable
     */
  timeout?: number | null;
}

export interface ExecuteResult {
  /** Standard output from the code */
  stdout: string;
  /** Standard error output */
  stderr: string;
  /** Exit code (0 = success) */
  exitCode: number;
  /** How long execution took in milliseconds */
  executionTimeMs: number;
  /**
     * Error message if execution failed
     * @nullable
     */
  error?: string | null;
}

export interface Session {
  id: number;
  name: string;
  code: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface SessionInput {
  /** @minLength 1 */
  name: string;
  code: string;
  description?: string;
}

export interface HistoryEntry {
  id: number;
  code: string;
  stdout: string;
  stderr: string;
  exitCode: number;
  executionTimeMs: number;
  executedAt: string;
}