# lib/api-zod/src/generated/types/executeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Node.js REPL API
 * OpenAPI spec version: 0.1.0
 */

export interface ExecuteInput {
  /** JavaScript/Node.js code to execute */
  code: string;
  /**
     * Execution timeout in milliseconds (default 10000)
     * @nullable
     */
  timeout?: number | null;
}