# lib/api-zod/src/generated/types/executeResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Node.js REPL API
 * OpenAPI spec version: 0.1.0
 */

export interface ExecuteResult {
  /** Standard output from the code */
  stdout: string;
  /** Standard error output */
  stderr: string;
  /** Exit code (0 = success) */
  exitCode: number;
  /** How long execution took in milliseconds */
  executionTimeMs: number;
  /**
     * Error message if execution failed
     * @nullable
     */
  error?: string | null;
}