# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Node.js REPL API
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryEntry {
  id: number;
  code: string;
  stdout: string;
  stderr: string;
  exitCode: number;
  executionTimeMs: number;
  executedAt: Date;
}