# lib/api-zod/src/generated/types/session.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Node.js REPL API
 * OpenAPI spec version: 0.1.0
 */

export interface Session {
  id: number;
  name: string;
  code: string;
  /** @nullable */
  description?: string | null;
  createdAt: Date;
  updatedAt: Date;
}