# lib/api-zod/src/generated/types/sessionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Node.js REPL API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionInput {
  /** @minLength 1 */
  name: string;
  code: string;
  description?: string;
}